requirements.txt
Flask==0.12.3
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
opencv-python
gdata==2.0.18
google-api-python-client==1.4.2
httplib2==0.9.2
itsdangerous==0.24
kombu==3.0.26
numpy==1.15.1
oauth2client==1.5.2
progressbar2==3.6.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pylibmc==1.5.0
pyserial==3.0
pytz==2015.4
requests==2.9.1
rsa==3.3
simplejson==3.8.1
six==1.10.0
socketIO-client==0.6.5
uritemplate==0.6
websocket-client==0.35.0
wsgiref==0.1.2
